Move `Signatures` to `dotty.tools.dotc.util`
And address review comments.
